scrutinizerの指摘を修正
